requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.0
flask-bcrypt==1.0.1
pandas==2.1.3
matplotlib==3.8.2
Flask-Mail==0.9.1
APScheduler==3.10.4
Werkzeug==2.3.7
python-dotenv==1.0.0
Jinja2==3.1.2
gunicorn==21.2.0
itsdangerous==2.1.2
SQLAlchemy==2.0.23
WTForms==3.1.1
